requirements.txt
PyMuPDF
python-docx
beautifulsoup4
python-dotenv
azure-ai-documentintelligence
openai